% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdata_validate.R
\name{spdata_validate}
\alias{spdata_validate}
\title{Validate the images/ component of a spatial data file}
\usage{
spdata_validate(path, type = "image", s3_client = NULL)
}
\arguments{
\item{path}{Path to the spatial data file to validate}

\item{s3_client}{An optional S3 client}
}
\value{
\code{TRUE} if the spatial data file is valid, otherwise an error is thrown
}
\description{
Validate the images/ component of a spatial data file
}
